Move talent grids to talent-grids.ts
